Refactor CLI args into separate file
